# backend/openapi.yaml
openapi: 3.1.0

info:
  title: Shellhacks Game API
  version: 0.1.0
  description: HTTP API for creating games.

    In local development the public base URL is `http://localhost:8080` (Caddy), which
    strips the `/api` prefix before proxying to the backend. Realtime gameplay is handled
    over Socket.IO and is documented in `asyncapi.yaml`.
  license:
    name: ISC
    identifier: ISC

security: []

servers:
  - url: "{scheme}://{host}"
    description: Local development (through Caddy; the `/api` prefix is stripped by the proxy)
    variables:
      scheme:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost:8080
        description: Host and port of the Caddy entrypoint.

tags:
  - name: games
    description: Game creation

paths:
  /api/games:
    post:
      tags:
        - games
      operationId: createGame
      summary: Create a new game
      description:
        Allocates a unique 6-character game code together with a host token.

        The `hostToken` is returned only by this endpoint. The host must present it as the
        `token` field of the Socket.IO handshake auth (see `asyncapi.yaml`) to be recognised
        as the host and be allowed to start the game.
      responses:
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGameResponse"
        "400":
          description: Malformed JSON request body
        "500":
          description: Unable to allocate a unique game code (unexpected server error)
  /api/games/{code}:
    get:
      tags:
        - games
      operationId: getGameStatus
      summary: Look up a game's status
      description: |
        Returns the current status of a game without joining it. The host uses this to decide
        whether a persisted game code is still a usable room (`lobby` or `active`) or should be
        replaced (`finished`, or `404` after the in-memory store was reset). Game codes are shared
        publicly, so this endpoint is unauthenticated.
      parameters:
        - name: code
          in: path
          required: true
          description: Six-character game code (case-insensitive).
          schema:
            type: string
            pattern: "^[ABCDEFGHJKMNPQRSTUVWXYZ23456789]{6}$"
            examples:
              - ABC234
      responses:
        "200":
          description: Game found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStatusResponse"
        "404":
          description: No game exists for that code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameNotFoundResponse"

components:
  schemas:
    CreateGameResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - hostToken
      properties:
        code:
          type: string
          description: Join code shared with players.
          pattern: "^[ABCDEFGHJKMNPQRSTUVWXYZ23456789]{6}$"
          examples:
            - ABC234
        hostToken:
          type: string
          description: Secret host credential. Returned only at creation time.
          examples:
            - 5v_qSvibmq-Fp9hREs476vZgm3aLhJfm
    GameStatusResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - status
      properties:
        code:
          type: string
          pattern: "^[ABCDEFGHJKMNPQRSTUVWXYZ23456789]{6}$"
          examples:
            - ABC234
        status:
          type: string
          enum:
            - lobby
            - active
            - finished
          description: |
            `lobby` – waiting to start, players may join; `active` – the round is running;
            `finished` – the round ended and the code should be replaced.
    GameNotFoundResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          const: GAME_NOT_FOUND
          examples:
            - GAME_NOT_FOUND